+++ blockapps-rest/src/util/api.util.ts
@@ -20,6 +20,7 @@ const Endpoint = {
   USER: `${blocUrl}/users/:username`,
   FILL: `${blocUrl}/users/user/:address/fill`,
   CONTRACTS: `${blocUrl}/contracts`,
+  CONTRACTS_XABI: `${blocUrl}/contracts/:name/:address`,
   STATE: `${blocUrl}/contracts/:name/:address/state`,
   TXRESULTS: `${blocUrl}/transactions/results`,
   SEND: `${strato23Url}/transaction`,